requirements.txt
qdrant-client==1.9.2
requests==2.32.3
myMuPDF==1.24.7
openai==1.35.4
langchain==0.2.6
python-dotenv==1.0.1
